requirements.txt
xarm-python-sdk==1.17.3
